- Rename Validation class to Ensure.
